+++ distribution/lognormal/doc.go
@@ -1,0 +1,10 @@
+// Package lognormal implements a lognormal continuous random variable
+//
+// lognormal distribution is a probability distribution parameterized by
+// two parameters, mean (also called location) and standard deviation
+// (also called scale)
+//
+// References:
+//
+// https://en.wikipedia.org/wiki/Log-normal_distribution
+package lognormal